aplicando as regras de autorização no sistema

+++ app/Http/Controllers/Admin/PapelController.php
@@ -7,6 +7,8 @@ use App\Http\Controllers\Controller;
 use App\Papel;
 use App\Permissao;
 
+use Illuminate\Support\Facades\Gate;
+
 class PapelController extends Controller
 {
     /**
@@ -16,6 +18,10 @@ class PapelController extends Controller
     */
     public function index()
     {
+        if (Gate::denies('papel-view')) {
+            abort(403,"Não autorizado!");
+        }
+
         $registros = Papel::all();
         $caminhos = [
             ['url'=>'/admin','titulo'=>'Admin'],
@@ -26,6 +32,10 @@ class PapelController extends Controller
 
     public function permissao($id)
     {
+        if (Gate::denies('papel-edit')) {
+            abort(403,"Não autorizado!");
+        }
+        
         $papel = Papel::find($id);
         $permissao = Permissao::all();
         $caminhos = [
@@ -39,6 +49,10 @@ class PapelController extends Controller
 
     public function permissaoStore(Request $request, $id)
     {
+        if (Gate::denies('papel-edit')) {
+            abort(403,"Não autorizado!");
+        }
+
         $papel = Papel::find($id);
         $dados = $request->all();
         $permissao = Permissao::find($dados['permissao_id']);
@@ -49,6 +63,10 @@ class PapelController extends Controller
 
     public function permissaoDestroy($id, $permissao_id)
     {
+        if (Gate::denies('papel-edit')) {
+            abort(403,"Não autorizado!");
+        }
+
         $papel = Papel::find($id);
         $permissao = Permissao::find($permissao_id);
 
@@ -63,6 +81,10 @@ class PapelController extends Controller
     */
     public function create()
     {
+        if (Gate::denies('papel-create')) {
+            abort(403,"Não autorizado!");
+        }
+
         $caminhos = [
             ['url'=>'/admin','titulo'=>'Admin'],
             ['url'=>route('papeis.index'),'titulo'=>'Papéis'],
@@ -80,13 +102,17 @@ class PapelController extends Controller
     */
     public function store(Request $request)
     {
-    if($request['nome'] && $request['nome'] != "Admin"){
-        Papel::create($request->all());
+        if (Gate::denies('papel-create')) {
+            abort(403,"Não autorizado!");
+        }
 
-        return redirect()->route('papeis.index');
-    }
+        if($request['nome'] && $request['nome'] != "Admin"){
+            Papel::create($request->all());
 
-    return redirect()->back();
+            return redirect()->route('papeis.index');
+        }
+
+        return redirect()->back();
     }
 
     /**
@@ -108,6 +134,10 @@ class PapelController extends Controller
     */
     public function edit($id)
     {
+        if (Gate::denies('papel-edit')) {
+            abort(403,"Não autorizado!");
+        }
+
         if (Papel::find($id)->nome == "Admin") {
             return redirect()->route('papeis.index');
         }
@@ -132,6 +162,10 @@ class PapelController extends Controller
     */
     public function update(Request $request, $id)
     {
+        if (Gate::denies('papel-edit')) {
+            abort(403,"Não autorizado!");
+        }
+
         if (Papel::find($id)->nome == "Admin") {
             return redirect()->route('papeis.index');
         }
@@ -150,9 +184,14 @@ class PapelController extends Controller
     */
     public function destroy($id)
     {
+        if (Gate::denies('papel-delete')) {
+            abort(403,"Não autorizado!");
+        }
+
         if (Papel::find($id)->nome == "Admin") {
             return redirect()->route('papeis.index');
         }
+
         Papel::find($id)->delete();
         return redirect()->route('papeis.index');
     }
